Move async controllers into their own file
