+++ server/controllers/index.js
@@ -4,13 +4,8 @@ import type {Promo} from '../model/promo';
 import {PromoFactory} from '../model/promo';
 import {createPageConfig} from '../model/page-config';
 import {getArticleStubs, getArticle, getSeries} from '../services/wordpress';
-import {getForwardFill, getUnpublishedSeries} from '../model/series';
-import { getSeriesColor } from '../data/series';
 import {PromoListFactory} from '../model/promo-list';
 import {PaginationFactory} from '../model/pagination';
-import {createNumberedList} from '../model/numbered-list';
-import { getLatestTweets } from '../services/twitter';
-import { getLatestInstagramPosts } from '../services/instagram';
 
 const maxItemsPerPage = 32;
 
@@ -78,115 +73,6 @@ export const series = async(ctx, next) => {
   return next();
 };
 
-export const seriesNav = async(ctx, next) => {
-  const {id} = ctx.params;
-  const {current} = ctx.request.query;
-  const seriesResponse = await getSeries(id, 6, 1);
-  const series = seriesResponse ? getForwardFill(seriesResponse) : getUnpublishedSeries(id);
-
-  const color = getSeriesColor(id);
-  const promoList = PromoListFactory.fromSeries(series);
-  const items = promoList.items.toJS();
-  const image = items[0].image;
-  const seriesNavModel = createNumberedList({
-    name: promoList.name,
-    image: image,
-    items: items,
-    color: color
-  });
-
-  ctx.render('components/numbered-list/index', {
-    current,
-    model: seriesNavModel,
-    modifiers: ['horizontal', 'sticky'],
-    data: {
-      classes: ['js-series-nav'],
-      sliderId: `series-nav--${id}`
-    }
-  });
-
-  ctx.body = {
-    html: ctx.body
-  };
-
-  return next();
-};
-
-export const seriesTransporter = async(ctx, next) => {
-  const { id } = ctx.params;
-  const { current } = ctx.request.query;
-  const seriesResponse = await getSeries(id, 6, {page: 1});
-  const series = seriesResponse ? getForwardFill(seriesResponse) : getUnpublishedSeries(id);
-
-  const color = getSeriesColor(id);
-  const promoList = PromoListFactory.fromSeries(series);
-  const items = promoList.items.toJS();
-  const image = items[0].image;
-  const seriesNavModel = createNumberedList({
-    name: promoList.name,
-    image: image,
-    items: items,
-    color: color
-  });
-
-  ctx.render('components/numbered-list/index', {
-    current,
-    model: seriesNavModel,
-    modifiers: ['transporter'],
-    data: {
-      classes: ['js-numbered-list-transporter'],
-      sliderId: `transporter--${id}`
-    }
-  });
-
-  ctx.body = {
-    html: ctx.body
-  };
-
-  return next();
-};
-
-export const latestInstagramPosts = async(ctx, next) => {
-  const instagramPosts = await getLatestInstagramPosts(10);
-
-  ctx.render('components/social-media-block/index', {
-    model: {
-      posts: instagramPosts,
-      service: 'Instagram',
-      icon: 'social/instagram',
-      url: 'https://instagram.com/wellcomecollection',
-      handle: 'wellcomecollection'
-    }
-  });
-
-  ctx.body = {
-    html: ctx.body
-  };
-
-  return next();
-};
-
-export const latestTweets = async(ctx, next) => {
-  const count = ctx.params.count || 4;
-  const tweets = await getLatestTweets(count);
-
-  ctx.render('components/social-media-block/index', {
-    model: {
-      posts: tweets,
-      service: 'Twitter',
-      icon: 'social/twitter',
-      url: 'https://twitter.com/explorewellcome',
-      handle: 'explorewellcome'
-    }
-  });
-
-  ctx.body = {
-    html: ctx.body
-  };
-
-  return next();
-};
-
 export const explore = async(ctx, next) => {
   const articleStubs = await getArticleStubs(50);
   const grouped = articleStubs.data.groupBy(stub => stub.headline.indexOf('A drop in the ocean:') === 0);
@@ -237,8 +123,7 @@ export const explore = async(ctx, next) => {
     topPromo,
     second3Promos,
     next8Promos,
-    collectorsPromo,
-    latestTweets
+    collectorsPromo
   });
 
   return next();
