+++ server/views/pages/explore.njk
@@ -79,9 +79,5 @@
     {% component 'async', { endpoint: '/latest-tweets', component: 'social-media-block-twitter' } %}
   </div>
 
-  <div data-component-name="Social media block" data-component-id="explore-instagram">
-    {% component 'async', { endpoint: '/latest-instagram-posts', component: 'social-media-block-instagram' } %}
-  </div>
-
   {% componentV2 'promo', collectorsPromo, {}, {sizes: '100vw'} %}
 {% endblock %}
